another overhaul (#337)
* another overhaul

- added backend signing for a single metadata id
- some small refactoring
- extracted upload from the CLI (maybe will do the same with other methods)

* optimized logging in the CLI

- logging now has levels
- added more readme content

* eslint update
