+++ js/packages/cli/package.json
@@ -25,6 +25,7 @@
     "bn.js": "^5.2.0",
     "commander": "^8.1.0",
     "form-data": "^4.0.0",
+    "loglevel": "^1.7.1",
     "node-fetch": "^2.6.1"
   },
   "devDependencies": {
