+++ js/packages/cli/src/cli.ts
@@ -1,43 +1,20 @@
-#!/usr/bin/env node
+#!/usr/bin/env ts-node
 import * as fs from 'fs';
 import * as path from 'path';
-import FormData from 'form-data';
 import { program } from 'commander';
 import * as anchor from '@project-serum/anchor';
 import BN from 'bn.js';
-import { MintLayout, Token } from '@solana/spl-token';
 
-import {
-  chunks,
-  fromUTF8Array,
-  loadCache,
-  parsePrice,
-  saveCache,
-  upload,
-} from './helpers/various';
-import { Keypair, PublicKey, SystemProgram } from '@solana/web3.js';
-import { createAssociatedTokenAccountInstruction } from './helpers/instructions';
-import {
-  CACHE_PATH,
-  CONFIG_ARRAY_START,
-  CONFIG_LINE_SIZE,
-  EXTENSION_JSON,
-  EXTENSION_PNG,
-  PAYMENT_WALLET,
-  TOKEN_METADATA_PROGRAM_ID,
-  TOKEN_PROGRAM_ID,
-} from './helpers/constants';
-import { sendTransactionWithRetryWithKeypair } from './helpers/transactions';
-import {
-  createConfig,
-  getCandyMachineAddress,
-  getMasterEdition,
-  getMetadata,
-  getTokenWallet,
-  loadAnchorProgram,
-  loadWalletKey,
-} from './helpers/accounts';
+import { fromUTF8Array, parsePrice } from './helpers/various';
+import { PublicKey } from '@solana/web3.js';
+import { CACHE_PATH, CONFIG_ARRAY_START, CONFIG_LINE_SIZE, EXTENSION_JSON, EXTENSION_PNG, } from './helpers/constants';
+import { getCandyMachineAddress, loadAnchorProgram, loadWalletKey, } from './helpers/accounts';
 import { Config } from './types';
+import { upload } from './commands/upload';
+import { loadCache, saveCache } from './helpers/cache';
+import { mint } from "./commands/mint";
+import { signAllUnapprovedMetadata, signMetadata } from "./commands/sign";
+import log from 'loglevel';
 
 program.version('0.0.1');
 
@@ -45,8 +22,9 @@ if (!fs.existsSync(CACHE_PATH)) {
   fs.mkdirSync(CACHE_PATH);
 }
 
-program
-  .command('upload')
+log.setLevel(log.levels.INFO);
+
+programCommand('upload')
   .argument(
     '<directory>',
     'Directory containing images named from 0-n',
@@ -54,21 +32,9 @@ program
       return fs.readdirSync(`${val}`).map(file => path.join(val, file));
     },
   )
-  .option(
-    '-e, --env <string>',
-    'Solana cluster env name',
-    'devnet', //mainnet-beta, testnet, devnet
-  )
-  .option(
-    '-k, --keypair <path>',
-    `Solana wallet location`,
-    '--keypair not provided',
-  )
-  // .argument('[second]', 'integer argument', (val) => parseInt(val), 1000)
   .option('-n, --number <number>', 'Number of images to upload')
-  .option('-c, --cache-name <string>', 'Cache file name', 'temp')
   .action(async (files: string[], options, cmd) => {
-    const { number, keypair, env, cacheName } = cmd.opts();
+    const {number, keypair, env, cacheName} = cmd.opts();
     const parsedNumber = parseInt(number);
 
     const pngFileCount = files.filter(it => {
@@ -79,243 +45,30 @@ program
     }).length;
 
     if (pngFileCount !== jsonFileCount) {
-      throw new Error(
-        `number of png files (${pngFileCount}) is different than the number of json files (${jsonFileCount})`,
-      );
+      throw new Error(`number of png files (${pngFileCount}) is different than the number of json files (${jsonFileCount})`);
     }
 
     if (parsedNumber < pngFileCount) {
-      throw new Error(
-        `max number (${parsedNumber})cannot be smaller than the number of elements in the source folder (${pngFileCount})`,
-      );
+      throw new Error(`max number (${parsedNumber})cannot be smaller than the number of elements in the source folder (${pngFileCount})`);
     }
 
-    const savedContent = loadCache(cacheName, env);
-    const cacheContent = savedContent || {};
-
-    if (!cacheContent.program) {
-      cacheContent.program = {};
-    }
-
-    let existingInCache = [];
-    if (!cacheContent.items) {
-      cacheContent.items = {};
-    } else {
-      existingInCache = Object.keys(cacheContent.items);
-    }
-
-    const seen = {};
-    const newFiles = [];
-
-    files.forEach(f => {
-      if (!seen[f.replace(EXTENSION_PNG, '').split('/').pop()]) {
-        seen[f.replace(EXTENSION_PNG, '').split('/').pop()] = true;
-        newFiles.push(f);
-      }
-    });
-    existingInCache.forEach(f => {
-      if (!seen[f]) {
-        seen[f] = true;
-        newFiles.push(f + '.png');
-      }
-    });
-
-    const images = newFiles.filter(val => path.extname(val) === EXTENSION_PNG);
-    const SIZE = images.length;
-
-    const walletKeyPair = loadWalletKey(keypair);
-    const anchorProgram = await loadAnchorProgram(walletKeyPair, env);
-
-    let config = cacheContent.program.config
-      ? new PublicKey(cacheContent.program.config)
-      : undefined;
-
-    for (let i = 0; i < SIZE; i++) {
-      const image = images[i];
-      const imageName = path.basename(image);
-      const index = imageName.replace(EXTENSION_PNG, '');
-
-      console.log(`Processing file: ${index}`);
-
-      const storageCost = 10;
-
-      let link = cacheContent?.items?.[index]?.link;
-      if (!link || !cacheContent.program.uuid) {
-        const manifestPath = image.replace(EXTENSION_PNG, '.json');
-        const manifestContent = fs
-          .readFileSync(manifestPath)
-          .toString()
-          .replace(imageName, 'image.png')
-          .replace(imageName, 'image.png');
-        const manifest = JSON.parse(manifestContent);
-
-        const manifestBuffer = Buffer.from(JSON.stringify(manifest));
-
-        if (i === 0 && !cacheContent.program.uuid) {
-          // initialize config
-          console.log(`initializing config`);
-          try {
-            const res = await createConfig(anchorProgram, walletKeyPair, {
-              maxNumberOfLines: new BN(parsedNumber || SIZE),
-              symbol: manifest.symbol,
-              sellerFeeBasisPoints: manifest.seller_fee_basis_points,
-              isMutable: true,
-              maxSupply: new BN(0),
-              retainAuthority: true,
-              creators: manifest.properties.creators.map(creator => {
-                return {
-                  address: new PublicKey(creator.address),
-                  verified: true,
-                  share: creator.share,
-                };
-              }),
-            });
-            cacheContent.program.uuid = res.uuid;
-            cacheContent.program.config = res.config.toBase58();
-            config = res.config;
-
-            console.log(
-              `initialized config for a candy machine with uuid: ${res.uuid}`,
-            );
-
-            saveCache(cacheName, env, cacheContent);
-          } catch (exx) {
-            console.error('Error deploying config to Solana network.', exx);
-            // console.error(exx);
-          }
-        }
-
-        if (!link) {
-          const instructions = [
-            anchor.web3.SystemProgram.transfer({
-              fromPubkey: walletKeyPair.publicKey,
-              toPubkey: PAYMENT_WALLET,
-              lamports: storageCost,
-            }),
-          ];
-
-          const tx = await sendTransactionWithRetryWithKeypair(
-            anchorProgram.provider.connection,
-            walletKeyPair,
-            instructions,
-            [],
-            'single',
-          );
-          console.info('transaction for arweave payment:', tx);
-
-          // data.append('tags', JSON.stringify(tags));
-          // payment transaction
-          const data = new FormData();
-          data.append('transaction', tx['txid']);
-          data.append('env', env);
-          data.append('file[]', fs.createReadStream(image), {
-            filename: `image.png`,
-            contentType: 'image/png',
-          });
-          data.append('file[]', manifestBuffer, 'metadata.json');
-          try {
-            const result = await upload(data, manifest, index);
-
-            const metadataFile = result.messages?.find(
-              m => m.filename === 'manifest.json',
-            );
-            if (metadataFile?.transactionId) {
-              link = `https://arweave.net/${metadataFile.transactionId}`;
-              console.log(`File uploaded: ${link}`);
-            }
-            console.log('setting cache for ', index);
-            cacheContent.items[index] = {
-              link,
-              name: manifest.name,
-              onChain: false,
-            };
-            saveCache(cacheName, env, cacheContent);
-          } catch (er) {
-            console.error(`Error uploading file ${index}`, er);
-          }
-        }
+    const startMs = Date.now();
+    log.info("started at: " + startMs.toString())
+    for (; ;) {
+      const successful = await upload(files, cacheName, env, keypair, parsedNumber);
+      if (successful) {
+        break;
+      } else {
+        log.warn("upload was not successful, rerunning");
       }
     }
+    const endMs = Date.now();
+    const timeTaken = new Date(endMs - startMs).toISOString().substr(11, 8);
+    log.info(`ended at: ${new Date(endMs).toString()}. time taken: ${timeTaken}`)
 
-    let updateSuccessful = true;
-    const keys = Object.keys(cacheContent.items);
-    try {
-      await Promise.all(
-        chunks(Array.from(Array(keys.length).keys()), 1000).map(
-          async allIndexesInSlice => {
-            for (
-              let offset = 0;
-              offset < allIndexesInSlice.length;
-              offset += 10
-            ) {
-              const indexes = allIndexesInSlice.slice(offset, offset + 10);
-              const onChain = indexes.filter(i => {
-                const index = keys[i];
-                return cacheContent.items[index]?.onChain || false;
-              });
-              const ind = keys[indexes[0]];
-
-              if (onChain.length != indexes.length) {
-                console.log(
-                  `Writing indices ${ind}-${keys[indexes[indexes.length - 1]]}`,
-                );
-                try {
-                  await anchorProgram.rpc.addConfigLines(
-                    ind,
-                    indexes.map(i => ({
-                      uri: cacheContent.items[keys[i]].link,
-                      name: cacheContent.items[keys[i]].name,
-                    })),
-                    {
-                      accounts: {
-                        config,
-                        authority: walletKeyPair.publicKey,
-                      },
-                      signers: [walletKeyPair],
-                    },
-                  );
-                  indexes.forEach(i => {
-                    cacheContent.items[keys[i]] = {
-                      ...cacheContent.items[keys[i]],
-                      onChain: true,
-                    };
-                  });
-                  saveCache(cacheName, env, cacheContent);
-                } catch (e) {
-                  console.log(
-                    `saving config line ${ind}-${
-                      keys[indexes[indexes.length - 1]]
-                    } failed`,
-                    e,
-                  );
-                  updateSuccessful = false;
-                }
-              }
-            }
-          },
-        ),
-      );
-    } catch (e) {
-      console.error(e);
-    } finally {
-      saveCache(cacheName, env, cacheContent);
-    }
-    console.log(`Done. Successful = ${updateSuccessful}. If 'false' - rerun`);
   });
 
-program
-  .command('verify')
-  .option(
-    '-e, --env <string>',
-    'Solana cluster env name',
-    'devnet', //mainnet-beta, testnet, devnet
-  )
-  .option(
-    '-k, --keypair <path>',
-    `Solana wallet location`,
-    '--keypair not provided',
-  )
-  .option('-c, --cache-name <string>', 'Cache file name', 'temp')
+programCommand('verify')
   .action(async (directory, cmd) => {
     const { env, keypair, cacheName } = cmd.opts();
 
@@ -331,7 +84,7 @@ program
 
     const keys = Object.keys(cacheContent.items);
     for (let i = 0; i < keys.length; i++) {
-      console.log('Looking at key ', i);
+      log.debug('Looking at key ', i);
       const key = keys[i];
       const thisSlice = config.data.slice(
         CONFIG_ARRAY_START + 4 + CONFIG_LINE_SIZE * i,
@@ -341,22 +94,16 @@ program
       const uri = fromUTF8Array([...thisSlice.slice(40, 240)]);
       const cacheItem = cacheContent.items[key];
       if (!name.match(cacheItem.name) || !uri.match(cacheItem.link)) {
-        console.log(
-          'Name',
-          name,
-          'or uri',
-          uri,
-          'didnt match cache values of',
-          cacheItem.name,
-          'and',
-          cacheItem.link,
-          ' marking to rerun for image',
-          key,
-        );
+        //leaving here for debugging reasons, but it's pretty useless. if the first upload fails - all others are wrong
+        // console.log(
+        //   `Name (${name}) or uri (${uri}) didnt match cache values of (${cacheItem.name})` +
+        //   `and (${cacheItem.link}). marking to rerun for image`,
+        //   key,
+        // );
         cacheItem.onChain = false;
         allGood = false;
       } else {
-        console.debug('Name', name, 'with', uri, 'checked out');
+        log.debug('Name', name, 'with', uri, 'checked out');
       }
     }
 
@@ -374,7 +121,7 @@ program
 
     const lineCount = new BN(config.data.slice(247, 247 + 4), undefined, 'le');
 
-    console.log(
+    log.info(
       `uploaded (${lineCount.toNumber()}) out of (${
         configData.data.maxNumberOfLines
       })`,
@@ -386,25 +133,13 @@ program
         }) is smaller than the uploaded one (${lineCount.toNumber()})`,
       );
     } else {
-      console.log('ready to deploy!');
+      log.info('ready to deploy!');
     }
 
     saveCache(cacheName, env, cacheContent);
   });
 
-program
-  .command('create_candy_machine')
-  .option(
-    '-e, --env <string>',
-    'Solana cluster env name',
-    'devnet', //mainnet-beta, testnet, devnet
-  )
-  .option(
-    '-k, --keypair <path>',
-    `Solana wallet location`,
-    '--keypair not provided',
-  )
-  .option('-c, --cache-name <string>', 'Cache file name', 'temp')
+programCommand('create_candy_machine')
   .option('-p, --price <string>', 'SOL price', '1')
   .action(async (directory, cmd) => {
     const { keypair, env, price, cacheName } = cmd.opts();
@@ -442,22 +177,11 @@ program
       },
     );
 
-    console.log(`create_candy_machine Done: ${candyMachine.toBase58()}`);
+    saveCache(cacheName, env, cacheContent);
+    log.info(`create_candy_machine finished. candy machine pubkey: ${candyMachine.toBase58()}`);
   });
 
-program
-  .command('set_start_date')
-  .option(
-    '-e, --env <string>',
-    'Solana cluster env name',
-    'devnet', //mainnet-beta, testnet, devnet
-  )
-  .option(
-    '-k, --keypair <path>',
-    `Solana wallet location`,
-    '--keypair not provided',
-  )
-  .option('-c, --cache-name <string>', 'Cache file name', 'temp')
+programCommand('set_start_date')
   .option('-d, --date <string>', 'timestamp - eg "04 Dec 1995 00:12:00 GMT"')
   .action(async (directory, cmd) => {
     const { keypair, env, date, cacheName } = cmd.opts();
@@ -468,8 +192,7 @@ program
     const walletKeyPair = loadWalletKey(keypair);
     const anchorProgram = await loadAnchorProgram(walletKeyPair, env);
 
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const [candyMachine, _] = await getCandyMachineAddress(
+    const [candyMachine] = await getCandyMachineAddress(
       new PublicKey(cacheContent.program.config),
       cacheContent.program.uuid,
     );
@@ -484,103 +207,71 @@ program
       },
     );
 
-    console.log('set_start_date Done', secondsSinceEpoch, tx);
+    log.info('set_start_date Done', secondsSinceEpoch, tx);
   });
 
-program
-  .command('mint_one_token')
-  .option(
-    '-e, --env <string>',
-    'Solana cluster env name',
-    'devnet', //mainnet-beta, testnet, devnet
-  )
-  .option(
-    '-k, --keypair <path>',
-    `Solana wallet location`,
-    '--keypair not provided',
-  )
-  .option('-c, --cache-name <string>', 'Cache file name', 'temp')
+programCommand('mint_one_token')
   .action(async (directory, cmd) => {
-    const { keypair, env, cacheName } = cmd.opts();
+    const {keypair, env, cacheName} = cmd.opts();
 
     const cacheContent = loadCache(cacheName, env);
-    const mint = Keypair.generate();
-
-    const walletKeyPair = loadWalletKey(keypair);
-    const anchorProgram = await loadAnchorProgram(walletKeyPair, env);
-    const userTokenAccountAddress = await getTokenWallet(
-      walletKeyPair.publicKey,
-      mint.publicKey,
-    );
-
     const configAddress = new PublicKey(cacheContent.program.config);
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const [candyMachineAddress, bump] = await getCandyMachineAddress(
-      configAddress,
-      cacheContent.program.uuid,
-    );
-    const candyMachine = await anchorProgram.account.candyMachine.fetch(
-      candyMachineAddress,
-    );
-    const metadataAddress = await getMetadata(mint.publicKey);
-    const masterEdition = await getMasterEdition(mint.publicKey);
-    const tx = await anchorProgram.rpc.mintNft({
-      accounts: {
-        config: configAddress,
-        candyMachine: candyMachineAddress,
-        payer: walletKeyPair.publicKey,
-        //@ts-ignore
-        wallet: candyMachine.wallet,
-        mint: mint.publicKey,
-        metadata: metadataAddress,
-        masterEdition,
-        mintAuthority: walletKeyPair.publicKey,
-        updateAuthority: walletKeyPair.publicKey,
-        tokenMetadataProgram: TOKEN_METADATA_PROGRAM_ID,
-        tokenProgram: TOKEN_PROGRAM_ID,
-        systemProgram: SystemProgram.programId,
-        rent: anchor.web3.SYSVAR_RENT_PUBKEY,
-        clock: anchor.web3.SYSVAR_CLOCK_PUBKEY,
-      },
-      signers: [mint, walletKeyPair],
-      instructions: [
-        anchor.web3.SystemProgram.createAccount({
-          fromPubkey: walletKeyPair.publicKey,
-          newAccountPubkey: mint.publicKey,
-          space: MintLayout.span,
-          lamports:
-            await anchorProgram.provider.connection.getMinimumBalanceForRentExemption(
-              MintLayout.span,
-            ),
-          programId: TOKEN_PROGRAM_ID,
-        }),
-        Token.createInitMintInstruction(
-          TOKEN_PROGRAM_ID,
-          mint.publicKey,
-          0,
-          walletKeyPair.publicKey,
-          walletKeyPair.publicKey,
-        ),
-        createAssociatedTokenAccountInstruction(
-          userTokenAccountAddress,
-          walletKeyPair.publicKey,
-          walletKeyPair.publicKey,
-          mint.publicKey,
-        ),
-        Token.createMintToInstruction(
-          TOKEN_PROGRAM_ID,
-          mint.publicKey,
-          userTokenAccountAddress,
-          walletKeyPair.publicKey,
-          [],
-          1,
-        ),
-      ],
-    });
+    const tx = await mint(keypair, env, configAddress);
 
-    console.log('Done', tx);
+    log.info('Done', tx);
   });
 
-program.command('find-wallets').action(() => {});
+programCommand('sign')
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  .option('-m, --metadata <string>', 'base58 metadata account id')
+  .action(async (directory, cmd) => {
+    const {keypair, env, metadata} = cmd.opts();
+
+    await signMetadata(
+      metadata,
+      keypair,
+      env
+    );
+  });
+
+programCommand('sign_all')
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  .action(async (directory, cmd) => {
+    const {keypair, env} = cmd.opts();
+
+    await signAllUnapprovedMetadata(
+      keypair,
+      env
+    );
+  });
+
+function programCommand(name: string) {
+  return program
+    .command(name)
+    .option(
+      '-e, --env <string>',
+      'Solana cluster env name',
+      'devnet', //mainnet-beta, testnet, devnet
+    )
+    .option(
+      '-k, --keypair <path>',
+      `Solana wallet location`,
+      '--keypair not provided',
+    )
+    .option('-l, --log-level <string>', 'log level', setLogLevel)
+    .option('-c, --cache-name <string>', 'Cache file name', 'temp');
+}
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+function setLogLevel(value, prev) {
+  if (value === undefined || value === null){
+    return
+  }
+  log.info("setting the log value to: " + value);
+  log.setLevel(value);
+}
+
+program.command('find-wallets').action(() => {
+});
 
 program.parse(process.argv);
