+++ js/packages/cli/src/helpers/accounts.ts
@@ -1,4 +1,4 @@
-import {Keypair, PublicKey, SystemProgram} from '@solana/web3.js';
+import { Keypair, PublicKey, SystemProgram } from '@solana/web3.js';
 import {
   CANDY_MACHINE,
   CANDY_MACHINE_PROGRAM_ID,
@@ -9,7 +9,9 @@ import {
 import * as anchor from '@project-serum/anchor';
 import fs from 'fs';
 import BN from "bn.js";
-import {createConfigAccount} from "./instructions";
+import { createConfigAccount } from "./instructions";
+import { web3 } from "@project-serum/anchor";
+import log from "loglevel";
 
 export const createConfig = async function (
   anchorProgram: anchor.Program,
@@ -29,7 +31,7 @@ export const createConfig = async function (
   },
 ) {
   const configAccount = Keypair.generate();
-  const uuid = configAccount.publicKey.toBase58().slice(0, 6);
+  const uuid = uuidFromConfigPubkey(configAccount.publicKey);
 
   return {
     config: configAccount.publicKey,
@@ -61,6 +63,10 @@ export const createConfig = async function (
   };
 };
 
+export function uuidFromConfigPubkey(configAccount: PublicKey) {
+  return configAccount.toBase58().slice(0, 6);
+}
+
 export const getTokenWallet = async function (
   wallet: PublicKey,
   mint: PublicKey,
@@ -76,7 +82,7 @@ export const getTokenWallet = async function (
 export const getCandyMachineAddress = async (
   config: anchor.web3.PublicKey,
   uuid: string,
-) => {
+): Promise<[PublicKey, number]> => {
   return await anchor.web3.PublicKey.findProgramAddress(
     [Buffer.from(CANDY_MACHINE), config.toBuffer(), Buffer.from(uuid)],
     CANDY_MACHINE_PROGRAM_ID,
@@ -86,7 +92,7 @@ export const getCandyMachineAddress = async (
 export const getConfig = async (
   authority: anchor.web3.PublicKey,
   uuid: string,
-) => {
+): Promise<[PublicKey, number]> => {
   return await anchor.web3.PublicKey.findProgramAddress(
     [Buffer.from(CANDY_MACHINE), authority.toBuffer(), Buffer.from(uuid)],
     CANDY_MACHINE_PROGRAM_ID,
@@ -125,20 +131,23 @@ export const getMasterEdition = async (
 };
 
 export function loadWalletKey(keypair): Keypair {
-  return Keypair.fromSecretKey(
+  const loaded = Keypair.fromSecretKey(
     new Uint8Array(JSON.parse(fs.readFileSync(keypair).toString())),
   );
+  log.info(`wallet public key: ${loaded.publicKey}`)
+  return loaded;
 }
 
 export async function loadAnchorProgram(walletKeyPair: Keypair, env: string) {
-  const solConnection = new anchor.web3.Connection(
-    `https://api.${env}.solana.com/`,
-  );
+  // @ts-ignore
+  const solConnection = new web3.Connection(web3.clusterApiUrl(env));
   const walletWrapper = new anchor.Wallet(walletKeyPair);
   const provider = new anchor.Provider(solConnection, walletWrapper, {
     preflightCommitment: 'recent',
   });
   const idl = await anchor.Program.fetchIdl(CANDY_MACHINE_PROGRAM_ID, provider);
 
-  return new anchor.Program(idl, CANDY_MACHINE_PROGRAM_ID, provider);
+  const program = new anchor.Program(idl, CANDY_MACHINE_PROGRAM_ID, provider);
+  log.info("program id from anchor", program.programId.toBase58());
+  return program;
 }
