+++ js/packages/cli/src/helpers/transactions.ts
@@ -13,6 +13,7 @@ import {
 } from '@solana/web3.js';
 import { getUnixTs, sleep } from './various';
 import { DEFAULT_TIMEOUT } from './constants';
+import log from "loglevel";
 
 interface BlockhashAndFeeCalculator {
   blockhash: Blockhash;
@@ -85,7 +86,7 @@ export async function sendSignedTransaction({
     },
   );
 
-  console.log('Started awaiting confirmation for', txid);
+  log.debug('Started awaiting confirmation for', txid);
 
   let done = false;
   (async () => {
@@ -109,13 +110,13 @@ export async function sendSignedTransaction({
       throw new Error('Timed out awaiting confirmation on transaction');
 
     if (confirmation.err) {
-      console.error(confirmation.err);
+      log.error(confirmation.err);
       throw new Error('Transaction failed: Custom instruction error');
     }
 
     slot = confirmation?.slot || 0;
   } catch (err) {
-    console.error('Timeout Error caught', err);
+    log.error('Timeout Error caught', err);
     if (err.timeout) {
       throw new Error('Timed out awaiting confirmation on transaction');
     }
@@ -125,7 +126,7 @@ export async function sendSignedTransaction({
         await simulateTransaction(connection, signedTransaction, 'single')
       ).value;
     } catch (e) {
-      console.error('Simulate Transaction error', e);
+      log.error('Simulate Transaction error', e);
     }
     if (simulateResult && simulateResult.err) {
       if (simulateResult.logs) {
@@ -145,7 +146,7 @@ export async function sendSignedTransaction({
     done = true;
   }
 
-  console.log('Latency', txid, getUnixTs() - startTime);
+  log.debug('Latency', txid, getUnixTs() - startTime);
   return { txid, slot };
 }
 
@@ -196,7 +197,7 @@ async function awaitTransactionSignatureConfirmation(
         return;
       }
       done = true;
-      console.log('Rejecting for timeout...');
+      log.warn('Rejecting for timeout...');
       reject({ timeout: true });
     }, timeout);
     try {
@@ -210,10 +211,10 @@ async function awaitTransactionSignatureConfirmation(
             confirmations: 0,
           };
           if (result.err) {
-            console.log('Rejected via websocket', result.err);
+            log.warn('Rejected via websocket', result.err);
             reject(status);
           } else {
-            console.log('Resolved via websocket', result);
+            log.debug('Resolved via websocket', result);
             resolve(status);
           }
         },
@@ -221,7 +222,7 @@ async function awaitTransactionSignatureConfirmation(
       );
     } catch (e) {
       done = true;
-      console.error('WS error in setup', txid, e);
+      log.error('WS error in setup', txid, e);
     }
     while (!done && queryStatus) {
       // eslint-disable-next-line no-loop-func
@@ -233,22 +234,22 @@ async function awaitTransactionSignatureConfirmation(
           status = signatureStatuses && signatureStatuses.value[0];
           if (!done) {
             if (!status) {
-              console.log('REST null result for', txid, status);
+              log.debug('REST null result for', txid, status);
             } else if (status.err) {
-              console.log('REST error for', txid, status);
+              log.error('REST error for', txid, status);
               done = true;
               reject(status.err);
             } else if (!status.confirmations) {
-              console.log('REST no confirmations for', txid, status);
+              log.error('REST no confirmations for', txid, status);
             } else {
-              console.log('REST confirmation for', txid, status);
+              log.info('REST confirmation for', txid, status);
               done = true;
               resolve(status);
             }
           }
         } catch (e) {
           if (!done) {
-            console.log('REST connection error: txid', txid, e);
+            log.error('REST connection error: txid', txid, e);
           }
         }
       })();
@@ -260,6 +261,6 @@ async function awaitTransactionSignatureConfirmation(
   if (connection._signatureSubscriptions[subId])
     connection.removeSignatureListener(subId);
   done = true;
-  console.log('Returning status', status);
+  log.debug('Returning status', status);
   return status;
 }
