+++ js/packages/cli/src/helpers/various.ts
@@ -1,9 +1,4 @@
 import { LAMPORTS_PER_SOL } from '@solana/web3.js';
-import { CACHE_PATH } from './constants';
-import path from 'path';
-import fs from 'fs';
-import FormData from 'form-data';
-import fetch from 'node-fetch';
 
 export const getUnixTs = () => {
   return new Date().getTime() / 1000;
@@ -53,41 +48,6 @@ export function fromUTF8Array(data: number[]) {
   return str;
 }
 
-export function chunks(array, size) {
-  return Array.apply(0, new Array(Math.ceil(array.length / size))).map(
-    (_, index) => array.slice(index * size, (index + 1) * size),
-  );
-}
-
-export function cachePath(env: string, cacheName: string) {
-  return path.join(CACHE_PATH, `${env}-${cacheName}`);
-}
-
-export function loadCache(cacheName: string, env: string) {
-  const path = cachePath(env, cacheName);
-  return fs.existsSync(path)
-    ? JSON.parse(fs.readFileSync(path).toString())
-    : undefined;
-}
-
-export function saveCache(cacheName: string, env: string, cacheContent) {
-  fs.writeFileSync(cachePath(env, cacheName), JSON.stringify(cacheContent));
-}
-
-export function parsePrice(price): number {
+export function parsePrice(price) {
   return Math.ceil(parseFloat(price) * LAMPORTS_PER_SOL);
-}
-
-export async function upload(data: FormData, manifest, index) {
-  console.log(`trying to upload ${index}.png: ${manifest.name}`);
-  return await (
-    await fetch(
-      'https://us-central1-principal-lane-200702.cloudfunctions.net/uploadFile4',
-      {
-        method: 'POST',
-        // @ts-ignore
-        body: data,
-      },
-    )
-  ).json();
 }
