+++ js/packages/cli/src/types.ts
@@ -1,13 +1,13 @@
-import {BN} from "@project-serum/anchor";
-import {PublicKey} from "@solana/web3.js";
+import { BN } from "@project-serum/anchor";
+import { PublicKey } from "@solana/web3.js";
 
 export class Creator {
-  address: string;
+  address: PublicKey;
   verified: boolean;
   share: number;
 
   constructor(args: {
-    address: string;
+    address: PublicKey;
     verified: boolean;
     share: number;
   }) {
@@ -32,6 +32,7 @@ export class ConfigData {
   isMutable: boolean;
   maxSupply: BN;
   retainAuthority: boolean;
+
   constructor(args: {
     name: string;
     symbol: string;
